+++ package.json
@@ -5,7 +5,7 @@
   "main": "./src/server.js",
   "scripts": {
     "start": "nodemon ./src/server",
-    "dev": "nodemon ./src/server",
+    "dev": "set NODE_ENV=production && nodemon ./src/server",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
     "prepare": "husky install",
